requirements.txt
Flask==2.1.1
grpcio==1.45.0
grpcio-tools==1.45.0